Modalarizing the rest of ui-utils
