+++ app/common/filters/trust.js
@@ -1,0 +1,9 @@
+'use strict';
+
+export default function($sce) {
+    'ngInject';
+
+    return function (text) {
+        return $sce.trustAsHtml(text);
+    };
+}
