+++ app/common/services/translate.js
@@ -1,0 +1,9 @@
+'use strict';
+
+export default function (gettextCatalog) {
+    'ngInject';
+
+    return function (str) {
+        return gettextCatalog.getString(str);
+    };
+}
